Refactor icon sizes to use UiIconSize enum
Replaced `UiIconSizes` constants with a `UiIconSize` enum for improved maintainability and clarity. Updated all relevant components to use the new enum and adjusted size handling logic accordingly. Also restructured `UiIcons` namespace for better organization.
