+++ src/Sidekick.Common.Blazor/Settings/Components/KeybindEditorInternal.razor
@@ -12,7 +12,7 @@
     @if (IsDuplicated)
     {
         <label class="mt-1 text-base font-medium text-amber-500 flex items-center flex-nowrap gap-2">
-            <Icon Svg="@UiIcons.Warning" Size="@UiIconSizes.Large" />
+            <Icon Svg="@UiIcons.Warning" Size="@UiIconSize.Medium" />
             <span>@Resources["Keybind_Already_In_Use"]</span>
         </label>
     }
