+++ src/Sidekick.Common.Ui/Alerts/AlertError.razor
@@ -1,5 +1,5 @@
-<div class="flex flex-nowrap items-center w-full @BackgroundClass p-3">
-    <Icon Svg="@UiIcons.Error" Size="@UiIconSizes.Large" />
+﻿<div class="flex flex-nowrap items-center w-full @BackgroundClass p-3">
+    <Icon Svg="@UiIcons.Error" Size="@UiIconSize.Medium" />
     <div class="ml-3 grow">
         @ChildContent
     </div>
