+++ src/Sidekick.Common.Ui/Alerts/AlertInfo.razor
@@ -1,5 +1,5 @@
-<div class="flex flex-nowrap items-center w-full bg-violet-600 p-3 text-white">
-    <Icon Svg="@UiIcons.Warning" Size="@UiIconSizes.Large" />
+﻿<div class="flex flex-nowrap items-center w-full bg-violet-600 p-3 text-white">
+    <Icon Svg="@UiIcons.Warning" Size="@UiIconSize.Medium" />
     <div class="ml-3 grow">
         @ChildContent
     </div>
