+++ src/Sidekick.Common.Ui/App/AppClose.razor
@@ -5,7 +5,7 @@
         tabindex="-1"
         title="@Resources["Close"]"
         class="p-2 text-base font-medium tracking-wide text-white transition-colors duration-200 bg-red-900 hover:bg-red-700">
-    <Icon Svg="@UiIcons.Close" Size="@UiIconSizes.Large" />
+    <Icon Svg="@UiIcons.Close" Size="@UiIconSize.Medium" />
 </button>
 
 @inject ICurrentView CurrentView
