+++ src/Sidekick.Common.Ui/App/AppHome.razor
@@ -1,4 +1,4 @@
-
+﻿
 @if (new Uri(NavigationManager.Uri).PathAndQuery != "/home")
 {
     <button type="button"
@@ -8,7 +8,7 @@
             tabindex="-1"
             title="@Resources["Home"]"
             class="p-2 text-base font-medium tracking-wide text-white transition-colors duration-200 bg-violet-900 hover:bg-violet-700">
-        <Icon Svg="@UiIcons.Home" Size="@UiIconSizes.Large" />
+        <Icon Svg="@UiIcons.Home" Size="@UiIconSize.Medium" />
     </button>
 }
