+++ src/Sidekick.Common.Ui/App/AppSettings.razor
@@ -8,7 +8,7 @@
             tabindex="-1"
             title="Open Settings"
             class="p-2 text-base font-medium border-l tracking-wide text-white transition-colors duration-200 border-stone-700 bg-stone-950 hover:bg-stone-700 hover:border-stone-600">
-        <Icon Svg="@UiIcons.Settings" Size="@UiIconSizes.Large" />
+        <Icon Svg="@UiIcons.Settings" Size="@UiIconSize.Medium" />
     </button>
 
     @if (Visible)
@@ -24,7 +24,7 @@
                             tabindex="-1"
                             title="@Resources["All_Settings"]"
                             class="p-2 text-base font-medium tracking-wide text-white border-r transition-colors duration-200 border-stone-700 bg-stone-950 hover:bg-stone-700 hover:border-stone-600 flex items-center gap-1 leading-none">
-                        <Icon Svg="@UiIcons.Settings" Size="@UiIconSizes.Large" />
+                        <Icon Svg="@UiIcons.Settings" Size="@UiIconSize.Medium" />
                         <span>@Resources["All_Settings"]</span>
                     </button>
                     <button type="button"
@@ -34,7 +34,7 @@
                             tabindex="-1"
                             title="@Resources["Close"]"
                             class="p-2 text-base font-medium tracking-wide text-white transition-colors duration-200 border-stone-700 bg-stone-950 hover:bg-stone-700 hover:border-stone-600 flex items-center gap-1 leading-none">
-                        <Icon Svg="@UiIcons.ArrowForward" Size="@UiIconSizes.Large" />
+                        <Icon Svg="@UiIcons.ArrowForward" Size="@UiIconSize.Medium" />
                         <span>@Resources["Close"]</span>
                     </button>
                 </div>
