+++ src/Sidekick.Common.Ui/Errors/ErrorFullScreenBoundary.razor
@@ -10,11 +10,11 @@
             <div class="flex min-w-full min-h-full items-center justify-center text-center p-4 flex-col">
                 @if (context.IsCritical)
                 {
-                    <Icon Svg="@UiIcons.Error" Class="text-red-700" Size="@UiIconSizes.XLarge" />
+                    <Icon Svg="@UiIcons.Error" Class="text-red-700" Size="@UiIconSize.Large" />
                 }
                 else
                 {
-                    <Icon Svg="@UiIcons.Warning" Class="text-orange-700" Size="@UiIconSizes.XLarge" />
+                    <Icon Svg="@UiIcons.Warning" Class="text-orange-700" Size="@UiIconSize.Large" />
                 }
 
                 <div class="my-3">
