+++ src/Sidekick.Common.Ui/Icons/UiIconSize.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Common.Ui.Icons;
+
+public enum UiIconSize
+{
+    Small,
+    Medium,
+    Large,
+}
