+++ src/Sidekick.Common.Ui/Poe/Items/ItemPropertyText.razor
@@ -13,7 +13,7 @@
         <TooltipRight>
             <ChildContent>
                 <span class="text-[#7f7f7f] inline-block ml-1">
-                    <Icon Svg="@UiIcons.PlusBox" Size="@UiIconSizes.Small" />
+                    <Icon Svg="@UiIcons.PlusBox" Size="@UiIconSize.Small" />
                 </span>
             </ChildContent>
             <TooltipContent>
