--- src/Sidekick.Common.Ui/UiIconSizes.cs
@@ -1,9 +1,0 @@
-namespace Sidekick.Common.Ui;
-
-public struct UiIconSizes
-{
-    public const string Small = "size-3";
-    public const string Medium = "size-4";
-    public const string Large = "size-6";
-    public const string XLarge = "size-12";
-}
